+++ src/app/charts/ScatterChart.tsx
@@ -20,12 +20,10 @@ const defaultAxesLabelFont = {
 };
 const defaultLineStyle = {
     color: '#008aad',
-    // color: '#c95d15',
     lineWidth: 1,
     highlightColor: '#d2933f',
     highlightWidth: 3
 };
-const defaultPlotGridLines = {visible: true, color: 'rgba(210,147,63,0.35)'};
 
 interface Axes {
     xAxisGenerator: Axis<number | {valueOf(): number}>;
@@ -159,7 +157,6 @@ function ScatterChart(props: Props): JSX.Element {
     const margin = {...defaultMargin, ...props.margin};
     const axisStyle = {...defaultAxesStyle, ...props.axisStyle};
     const axisLabelFont = {...defaultAxesLabelFont, ...props.axisLabelFont};
-    const plotGridLines = {...defaultPlotGridLines, ...props.plotGridLines};
     const tooltip: TooltipStyle = {...defaultTooltipStyle, ...props.tooltip};
     const magnifier = {...defaultRadialMagnifierStyle, ...props.magnifier};
     const lineStyle = {...defaultLineStyle, ...props.lineStyle};
@@ -173,7 +170,6 @@ function ScatterChart(props: Props): JSX.Element {
     const mainGRef = useRef<Selection<SVGGElement, any, null, undefined>>();
 
     const magnifierRef = useRef<MagnifierSelection>();
-    // const magnifierAxisRef = useRef<Selection<BaseType, number, BaseType, any>>();
     const magnifierXAxisRef = useRef<LineSelection>();
     const magnifierXAxisLabelRef = useRef<Selection<SVGTextElement, any, SVGGElement, undefined>>();
     const magnifierYAxisRef = useRef<LineSelection>();
@@ -849,7 +845,7 @@ function ScatterChart(props: Props): JSX.Element {
     function magnifierLensAxisLabels(ticks: Array<number>, selection: GSelection): Selection<SVGTextElement, number, SVGGElement, any> {
         return selection
             .selectAll('text')
-            .data([-5, -1, 0, 1, 5])
+            .data(ticks)
             .enter()
             .append('text')
             .attr('fill', axisLabelFont.color)